勤怠編集　バック実装　須藤

+++ 2401teamdevelopment/src/main/java/com/example/demo/entity/AttendanceEditEntity.java
@@ -29,12 +29,12 @@ public class  AttendanceEditEntity implements Serializable {
   @Id
   @Column(name = "attendance_id")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long attendance_id;
+  private Integer  attendance_id;
   /**
    * ユーザーID
    */
   @Column(name = "user_id")
-  private Long user_id;
+  private Integer  user_id;
   /**
    * ステータス
    */
